player actor
some player actor code;

+++ actor/src/com/stone/actor/factory/IActorFactory.java
@@ -3,10 +3,28 @@ package com.stone.actor.factory;
 import com.stone.actor.manager.ManagerActor;
 import com.stone.actor.player.PlayerActor;
 
+/**
+ * the actor factory;
+ * 
+ * @author crazyjohn
+ *
+ */
 public interface IActorFactory {
 
-	public PlayerActor createPlayerActor(int playerId);
+	/**
+	 * create player actor;
+	 * 
+	 * @param playerId
+	 * @return
+	 */
+	public PlayerActor createPlayerActor(long playerId);
 
-	public ManagerActor createManagerActor(int managerId);
+	/**
+	 * create global manager actor;
+	 * 
+	 * @param managerId
+	 * @return
+	 */
+	public ManagerActor createManagerActor(long managerId);
 
 }
